requirements.txt
fastapi
uvicorn
faiss-cpu  # Or faiss-gpu if using GPU
numpy
pymongo
python-dotenv
pydantic
